refactor(cost): extract reconciliation repository
