+++ src/services/portfolio_transaction_processing_service/app/infrastructure/cost_basis/__init__.py
@@ -1,6 +1,13 @@
 """Expose organized infrastructure adapters and records for cost-basis processing."""
 
+from .corporate_action_reconciliation_repository import (
+    SqlAlchemyCorporateActionReconciliationRepository,
+)
 from .reference_data_repository import SqlAlchemyCostBasisReferenceDataRepository
 from .staged_effects import StagedCostEffects
 
-__all__ = ["SqlAlchemyCostBasisReferenceDataRepository", "StagedCostEffects"]
+__all__ = [
+    "SqlAlchemyCorporateActionReconciliationRepository",
+    "SqlAlchemyCostBasisReferenceDataRepository",
+    "StagedCostEffects",
+]
