+++ src/services/portfolio_transaction_processing_service/app/infrastructure/cost_processing_adapter.py
@@ -17,6 +17,7 @@ from ..application.cost_basis_processing import (
 from ..domain import BookedTransaction
 from ..domain.transaction import SettlementCashValidationError
 from ..ports import (
+    CorporateActionReconciliationRepository,
     CostBasisInstrumentReference,
     CostBasisPortfolioReference,
     CostBasisReferenceDataPort,
@@ -53,6 +54,7 @@ class CostEffectsStager(Protocol):
         portfolio: CostBasisPortfolioReference,
         instrument: CostBasisInstrumentReference | None,
         repo: CostCalculatorRepository,
+        reconciliation_repository: CorporateActionReconciliationRepository,
         cost_basis_method: CostBasisMethod,
         outbox_repo: OutboxRepository,
         correlation_id: str,
@@ -68,11 +70,13 @@ class CostProcessingCompatibilityAdapter:
         workflow: CostEffectsStager,
         repository: CostCalculatorRepository,
         reference_data: CostBasisReferenceDataPort,
+        reconciliation_repository: CorporateActionReconciliationRepository,
         outbox_repository: OutboxRepository,
     ) -> None:
         self._workflow = workflow
         self._repository = repository
         self._reference_data = reference_data
+        self._reconciliation_repository = reconciliation_repository
         self._outbox_repository = outbox_repository
 
     async def stage_event(
@@ -100,6 +104,7 @@ class CostProcessingCompatibilityAdapter:
             portfolio=portfolio,
             instrument=instrument,
             repo=self._repository,
+            reconciliation_repository=self._reconciliation_repository,
             cost_basis_method=prepared.cost_basis_method,
             outbox_repo=self._outbox_repository,
             correlation_id=correlation_id,
